Add files via upload
Old homeworks I mostly haven't touched in years; I can definitely write better code these days. The first is a simulation of managing processes for an operating system, requiring managing memory properly and comparing various fit algorithms. The second is an implementation of the perceptron with pocket algorithm to distinguish between 1 and 5 as written on pieces of mail in different handwriting.